update netbean generated entities class

+++ src/main/java/com/airconsole/wm_cms/model/repository/GroupPageRoleRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupPageRoleRepo extends JpaRepository<GroupPageRoleEntity, Long> {
+public interface GroupPageRoleRepo extends JpaRepository<GroupPageRoleEntity, Integer> {
 }
