+++ src/main/java/com/airconsole/wm_cms/model/repository/PageRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PageRepo extends JpaRepository<PageEntity, Long> {
+public interface PageRepo extends JpaRepository<PageEntity, Integer> {
 }
