+++ src/main/java/com/airconsole/wm_cms/model/repository/QuestionRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionRepo extends JpaRepository<QuestionEntity, Long> {
+public interface QuestionRepo extends JpaRepository<QuestionEntity, Integer> {
 }
